# Modernize test/perf/speed.jl