@IainNZ 2018-11-10
I set up a Julia 1.0/JuMP 0.18 environment so I can actually properly run these in parallel:

```
IAINMBP:JuMP18 idunning$ j1 --project=. ../dev/JuMP/test/perf/speed.jl
P-Median(100 facilities, 100 customers, 5000 locations) benchmark:
BenchmarkTools.Trial:
  memory estimate:  268.90 MiB
  allocs estimate:  4507937
  --------------
  minimum time:     234.328 ms (0.00% GC)
  median time:      552.879 ms (56.86% GC)
  mean time:        597.856 ms (59.84% GC)
  maximum time:     1.427 s (82.95% GC)
  --------------
  samples:          9
  evals/sample:     1
```